requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
requests==2.32.3
langchain==0.3.3
langchain-community==0.3.3
langchain-huggingface==0.1.0
pymupdf==1.24.12
sentence-transformers==3.2.1
chromadb==0.5.13
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9 